pyproject.toml
[tool.poetry]
name = "iscc-service"
version = "0.1.0"
description = "ISCC Web Service API"
authors = ["Titusz Pan <tp@py7.de>"]
readme = "README.md"
homepage = "https://iscc.codes/"
repository = "https://github.com/iscc/iscc-service"
keywords = ["iscc", "media", "identifier", "blockchain", "cli", "generator", "webservice", "api"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.47.1"
iscc-cli = "^0.9.1"
uvicorn = "^0.11.2"
bitstring = "^3.1.6"

[tool.poetry.dev-dependencies]
pytest = "^5.0"
black = "^19.10b0"

[build-system]
requires = ["poetry==1.0.0"]
build-backend = "poetry.masonry.api"
